perf: limit audit log search
`MemberUpdateEvent` fires when the User object updates as well, not just on nickname changes. If a non-nickname update happens, it will search through 90 days worth of audit logs without a `.take(n)` limit.

Limit is put on ban events as well in case it misses the ban audit log (e.g. delay was too short).